feat: mirai-hibernate-plugin
